Update command
Updates the remindme function with the "every" command
Reworks segment_command

+++ cogs/remindme.py
@@ -4,7 +4,8 @@ from datetime import datetime
 
 
 class remindme(commands.Cog, name="tatsu_is_bad"):
-    r_segment_command = re.compile(r"\s(.*\s)?in (.*)")  # Group 1 is message, group 2 is time.
+    r_interval = re.compile(r"(?:^)(.+?) yreve")  # Gets the interval, if it exists.
+    r_in_time = re.compile(r"(?:yreve |^)(.+?) ni")
 
     def __init__(self, bot):
         self.bot = bot
@@ -17,17 +18,21 @@ class remindme(commands.Cog, name="tatsu_is_bad"):
     async def remind(self, ctx):
         if not await self.bot.has_perm(ctx, dm=True): return
 
-        message, time = self.segment_command(ctx.message.content)
-        if not message:
+        reminder = self.segment_command(ctx.message.content)
+        if not reminder["time"] and not reminder["interval"]:
             await ctx.send("Provide me with a time to annoy you.")
             return
 
-        time = self.bot.time_from_string(time)  # Convert the time from the command into seconds.
+        time = self.bot.time_from_string(reminder["time"])  # Convert the time from the command into seconds.
         time = min(86400 * 30, time)  # Limit to 1 month.
         endtime = self.bot.time_from_now(seconds=time)
         to_remind = self.bot.admin_override(ctx)
 
-        self.bot.cursor.execute("INSERT INTO remindme VALUES(?,?,?)", [message, to_remind.id, endtime])
+        repeat_interval = self.bot.time_from_string(reminder["interval"])
+        if repeat_interval and repeat_interval < 300:  # Repeat must be at least 5 minutes if not 0.
+            repeat_interval = 300
+
+        self.bot.cursor.execute("INSERT INTO remindme VALUES(?,?,?,?)", [reminder["msg"], to_remind.id, endtime, repeat_interval])
         self.bot.cursor.execute("commit")
 
         how_long = self.bot.time_to_string(seconds=time)
@@ -40,25 +45,48 @@ class remindme(commands.Cog, name="tatsu_is_bad"):
         message = ""
 
         at_time = self.bot.get_variable(ctx.message.content, "at", type="keyword", default=False)
-        in_time = True # self.bot.get_variable(ctx.message.content, "in", type="keyword", default=False)
+        remove_id = int(self.bot.get_variable(ctx.message.content, "remove", type="int", default=0))
 
-        c = self.bot.cursor
-        c.execute("SELECT * FROM remindme WHERE user_id=?", [user])
+        # Return a list of reminders
+        if not remove_id:
+            c = self.bot.cursor
+            c.execute("SELECT rowid, * FROM remindme WHERE user_id=?", [user])
 
-        for entry in c.fetchall():
-            if at_time:
-                time = datetime.fromtimestamp(entry[2]).strftime("%Y-%m-%d %H:%M:%S GMT")
-                time = f"{time}"
-                message += f"{time}: {entry[0]}\n"
-            elif in_time:
-                time = self.bot.time_to_string(seconds=entry[2] - self.bot.time_now())
-                time = f" ------ in {time}"
-                message += f"{entry[0]}{time}\n"
+            for entry in c.fetchall():
+                if at_time:
+                    time = datetime.fromtimestamp(entry[3]).strftime("%Y-%m-%d %H:%M:%S GMT")
+                    time = f"{time}"
+                    message += f"{entry[0]}: {time}: {entry[1]}\n"
+                else:
+                    time = self.bot.time_to_string(seconds=entry[3] - self.bot.time_now())
+                    time = f" ------ in {time}"
+                    if entry[4]:
+                        reminder_amount = self.bot.time_to_string(seconds=entry[4])
+                        time += f" (every {reminder_amount})"
 
-        if not message:
-            message = "No reminders!"
+                    message += f"{entry[0]}: {entry[1]}{time}\n"
 
-        await ctx.send(self.bot.escape_message(message))
+            if not message:
+                message = "No reminders!"
+
+            await ctx.send(self.bot.escape_message(message))
+        # Remove reminder.
+        else:
+            # Check the user actually owns reminder emote.
+            self.bot.cursor.execute(f"SELECT user_id FROM remindme WHERE rowid={remove_id}")
+            result = self.bot.cursor.fetchone()
+            if result:
+                if user != result[0]:
+                    await ctx.send("You don't own this reminder.")
+                    return
+            else:
+                await ctx.send("No reminder with this id!")
+                return
+
+            self.bot.cursor.execute(f"DELETE FROM remindme WHERE rowid={remove_id}")
+            self.bot.cursor.execute("commit")
+            await ctx.send("Reminder deleted.")
+
 
     # Timed command
     async def remind_time(self, time_now):
@@ -73,21 +101,31 @@ class remindme(commands.Cog, name="tatsu_is_bad"):
                     print(f"Could not message {target[1]}")
                     pass
 
-                self.bot.cursor.execute(f"DELETE FROM remindme WHERE rowid={target[0]}")
+                if not target[4]:  # Repeat reminders.
+                    self.bot.cursor.execute(f"DELETE FROM remindme WHERE rowid={target[0]}")
+                else:
+                    new_time = target[4] + self.bot.time_now()
+                    self.bot.cursor.execute(f"UPDATE remindme SET time={new_time} WHERE rowid={target[0]}")
+
                 self.bot.cursor.execute("commit")
 
-    @commands.command(name="remind.help")
+    @commands.command(name="remind.help", aliases=["remindme.help"])
     async def remind_help(self, ctx):
         if not await self.bot.has_perm(ctx, dm=True): return
         docstring = """
         ```Set a reminder for yourself! Stolen from tatsu due to his unreliability.
         Accepted time units:
             second, minute, hour, day, week (and some others~)
+            
+        Arguments:
+            in - How long until the first reminder
+            every - How regularly to remind you of this. Omit for only once. Has to be after "in" argument.
         
         Examples:
             c.remind owo in 0.5 minutes
-            c.remind in 1 hour
-            c.remindme you're cool somewhere in there in 1 day```
+            c.remindme you're cool somewhere in there in 1 day
+            c.remind wake up time in 2 hours 10 minutes every 24 hours
+            c.remindme dummy thicc every 10 hours```
         """
         docstring = self.bot.remove_indentation(docstring)
         await ctx.send(docstring)
@@ -96,11 +134,15 @@ class remindme(commands.Cog, name="tatsu_is_bad"):
     async def reminders_help(self, ctx):
         if not await self.bot.has_perm(ctx, dm=True): return
         docstring = """
-        ```View your reminders!
-        Can be provided with the keyword "in" to change from absolute time to relative time.
+        ```Manage your reminders!
+        
+        Arguments:
+            remove - Delete the reminder with the provided ID. ID is the leftmost value in the reminders table.
+            in - Change display format for reminders.
         
         Examples:
             c.reminders
+            c.reminders remove=21
             c.reminders in```
         """
         docstring = self.bot.remove_indentation(docstring)
@@ -111,14 +153,39 @@ class remindme(commands.Cog, name="tatsu_is_bad"):
         Segments a command up into the reminder, and the timer.
         Returns [reminder, time] or [None, None] if not found.
         """
-        match = re.search(self.r_segment_command, command)
-        if not match:
-            return [None, None]
+        # Remove the invoking of the command.
+        command = command.replace("c.remind ", "")
+        command = command.replace("c.remindme ", "")
+        command = command[::-1]  # Reverse the string to match backwards.
 
-        m = match.group(1)
-        if not m:
-            m = "Reminder!"
-        return [m, match.group(2)]
+        # TODO: Clean this code, jeez.
+        match = {}
+        # Get the last matches.
+        match["interval"] = re.search(self.r_interval, command)
+        if match["interval"]:
+            start = match["interval"].start(1)
+            end = match["interval"].end(1)
+            match["interval"] = match["interval"].group(1)[::-1]
+            command = command[:start] + command[end + 6:]  # Remove this from the command.
+        else:
+            match["interval"] = ""
+
+        match["time"] = re.search(self.r_in_time, command)
+        if match["time"]:
+            start = match["time"].start(1)
+            end = match["time"].end(1)
+            match["time"] = match["time"].group(1)[::-1]  # Reverse the match to return it to normal
+            command = command[:start] + command[end + 3:]  # remove this from the command.
+        else:
+            match["time"] = ""
+
+        match["msg"] = command.strip()[::-1]  # Message is whatever remains of the command after all invoking info has been stripped.
+
+        # Default reminder message.
+        if not match["msg"]:
+            match["msg"] = "Reminder!"
+
+        return match
 
     def init_db(self, cursor):
         cursor.execute("begin")
@@ -126,7 +193,8 @@ class remindme(commands.Cog, name="tatsu_is_bad"):
             "CREATE TABLE IF NOT EXISTS remindme ("
             "message STRING,"  # What to remind them of.
             "user_id INTEGER,"  # Who to remind
-            "time INTEGER"  # When to remind them
+            "time INTEGER,"  # When to remind them
+            "interval INTEGER"  # For repeated reminders, how regularly to remind.
             ")")
         cursor.execute("commit")
